HashBlotto

Documentation for HashBlotto.

HashBlotto.HashBlotto

HashBlotto.all_random

HashBlotto.greed

HashBlotto.HashBlotto — Module

HashBlotto

A package that is build by the HashBlotto Team in C25. The project is based on the Google HashCode 2014 competition.

HashBlotto.all_random — Method

all_random(city)

Function that returns the best random solution after 10 iterations of the random solution and returns the best random solution of those 10.

HashBlotto.greed — Method

greed(City)

function that is greedy in it's approach to finding the best path, the logic it follows is looking at the distance/time ratio of the next junction and choosing the one with the highest ratio, if a car has already gone down one of the roads then we will calculate the reward based off of reward/ 2 * times it has gone down the road –––––––––––––––––––––––––––––––––––– Building off of the HashCode2014 package (https://github.com/gdalle/HashCode2014.jl) that was already built out for us, I definetly made it better as it is smarter in how it chooses paths than (https://github.com/gdalle/HashCode2014.jl/blob/main/src/random_walk.jl)