HashBlotto

Documentation for HashBlotto.

HashBlotto.HashBlotto

HashBlotto.Distributions

HashBlotto.all_random

HashBlotto.get_best_street

HashBlotto.get_value

HashBlotto.graph_structure

HashBlotto.greed

HashBlotto.HashBlotto — Module

HashBlotto

A package that is build by the HashBlotto Team in C25. The project is based on the Google HashCode 2014 competition.

HashBlotto.Distributions — Method

Distributions(city, min, max, step)

Shows the distributions of the distances of the city at different penalty values

HashBlotto.all_random — Method

all_random(city)

Iterates through 10 random walks, returning the best found solution

HashBlotto.get_best_street — Function

get_best_street(current_junction, street_candidates, visited)

Returns the best street that the google maps car can go to.

HashBlotto.get_value — Method

get_value (current_junction, visited)

Returns the value of the current junction. The current reward system acting under the reward equation:

val * penalty ^ visited[node] + (penalty) * adjacent_reward

HashBlotto.graph_structure — Method

graph_structure(city::City)

Given the city the function will return a precomupted adjacency dictionary. The keys of the dictionary will hold the the index of the vertice The values of the dictionary will be a list of tuples of (nextnodeindex, distance, duration, distance/duration)

HashBlotto.greed — Method

greed(City)

A greedy construction heuristic solution: Greedily maximize the distance/time ratio of the next junction. Reward is scaled down by 2 * repetitions of a street [if more than once] to discourage repeated edges

We improve on the prior work of the HashCode2014 package (https://github.com/gdalle/HashCode2014.jl) by designing and implementing a greedy heuristic (https://github.com/gdalle/HashCode2014.jl/blob/main/src/random_walk.jl)